requirements.txt
openai
python-dotenv
sounddevice
soundfile
numpy
pyttsx3
webrtcvad 